Add getResult method to SlotMachine interface
This commit introduces the getResult method to the SlotMachine interface and implements it in the BasicSlotMachine class. The new method provides access to the last result of a spin, enhancing the ability to retrieve spin outcomes easily. The AbstractSlotMachine class is also updated to include an abstract definition of this method.